refactor: share board generation
