+++ webapp/src/utils/generateBoard.js
@@ -1,0 +1,1 @@
+export { default } from '../../../bot/logic/generateBoard.js';
